+++ NEWS.md
@@ -1,3 +1,13 @@
+# mapSpain (development version)
+
+-   Changes on how to handle modifications on Canary Islands objects (#101):
+    -   Add a helper function for displace stand-alone **sf** objects in Canary
+        Islands: `esp_move_can()`.
+    -   `esp_move_can()` is used internally on all functions.
+-   Add a new function to show the current cache directory:
+    `esp_detect_cache_dir()`.
+-   `mapSpain::layer_spatraster()` removed.
+
 # mapSpain 0.8.0
 
 -   Improve download of NUTS data from **giscoR**.
@@ -87,7 +97,7 @@
 -   Move minimum version of `giscoR` to v0.2.4
 -   Fix typos on `esp_dict_translate()` #36.
 -   Not run examples on tiles, as the server sometimes doesn't respond.
--   Refactor `sysdata.rda`.
+-   Re factor `sysdata.rda`.
 -   CRAN fixes:
     -   Removed broken link on `addProviderEspTiles()`.
     -   Vignette removed (CRAN warning).